requirements.txt
Django>=5.0

# API (Login/Register endpoints)
djangorestframework

# JWT Token authentication
djangorestframework-simplejwt
PyJWT
bcrypt

# Database Drivers
pymysql
cryptography # It is required from pymysql in order to handle the authentication method used by MySQL

# Development tools
django-extensions
Werkzeug
pyOpenSSL